+++ app/views/providers/means/outstanding_mortgages/show.html.erb
@@ -1,7 +1,7 @@
 <%= form_with(
         model: @form,
         scope: :legal_aid_application,
-        url: providers_legal_aid_application_outstanding_mortgage_path,
+        url: providers_legal_aid_application_means_outstanding_mortgage_path,
         method: :patch,
         local: true) do |form| %>
     <%= page_template(page_title: t('.h1-heading'), template: :basic, form: form) do %>
